added 5490 dropdown changes and updated tests
